pyproject.toml
[project]
name = "hecm"
dynamic = ["version"]
description = "Epub reader for the 21st century"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "datasets>=4.1.1",
    "dspy>=3.0.3",
    "github3-py>=4.0.1",
    "gitpython>=3.1.45",
    "openai>=2.0.0",
    "python-dotenv>=1.1.1",
    "tqdm>=4.67.1",
    "weave>=0.52.8",
]

[project.scripts]
hecm = "hecm.run:run_hecm"

[dependency-groups]
dev = [
    "uv>=0.4.20",
    "pytest>=8.3.3",
    "isort>=5.13.2",
    "black>=24.10.0",
    "ruff>=0.6.9",
    "textual-dev>=1.7.0",
    "openai>=1.108.2",
    "pyright>=1.1.405",
    "pynvim>=0.6.0",
]
docs = [
    "mkdocs>=1.6.1",
    "mkdocstrings>=0.26.1",
    "mkdocstrings-python>=1.11.1",
    "mkdocs-material>=9.5.39",
    "mkdocs-minify-plugin>=0.8.0",
    "mkdocs-glightbox>=0.4.0",
    "mkdocs-jupyter>=0.25.0",
    "jupyter>=1.1.1",
]
build = [
    "twine>=5.1.1",
    "wheel>=0.40",
    "build>=0.10"
]

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.targets.wheel]
packages = ["hecm"]

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = ["I"]

[tool.ruff.lint.isort]
known-first-party = ["hecm"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]
combine-as-imports = true
